requirements.txt
openai-agents>=0.6.0
pydantic>=2.0
openai>=1.0
python-dotenv
click>=8.0
rich>=13.0
typing-extensions>=4.0